requirements.txt
redis
RPi.GPIO
lgpio
gpiozero 
paho-mqtt
python-dotenv